// TMS_SunYun/Client/src/app/cmc/model.ts
export class Messages {
    EventId: string = "";
    ReportTime: string = "";
    MessageCode: string = "";
    MessageBody: string = "";
    ChatId: string = "";
    ChatGroup: string = "";
    LmUser: string = "";
    LmTime: string = "";
}

export class SmsMessages {
    EventId: string = "";
    ReportTime: string = "";
    MessageCode: string = "";
    MessageBody: string = "";
    PhoneNo: string = "";
    LmUser: string = "";
    LmTime: string = "";
}

export class Users {
    UserName: string = '';
    UserId: string = '';
    Password: string = '123';
    ConfirmPassword: string = '123';
    OldPassword: string = '123';
    NewPassword: string = '123';
    Email: string = '';
    ContactNo: string = '';
    Role: string = '';
}